+++ src/mock/browser.ts
@@ -1,4 +1,4 @@
 import { setupWorker } from "msw/browser";
-import { handlers } from "./handlers.ts";
+import { handlers } from "./handlers/handlers.ts";
 
 export const worker = setupWorker(...handlers);
