+++ src/mock/handlers/handlers.ts
@@ -1,0 +1,3 @@
+import { cartItemsHandler } from "./cartItems";
+
+export const handlers = [...cartItemsHandler];
