+++ src/mock/server.ts
@@ -1,4 +1,4 @@
 import { setupServer } from "msw/node";
-import { handlers } from "./handlers";
+import { handlers } from "./handlers/handlers.ts";
 
 export const server = setupServer(...handlers);
